+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changelog
 * Clear all data will also clear local caches
 * Show parameters in usage data
 * Added restriction for deprecated *contacts/people* ([issue](/../../issues/1367))
+* Added restriction for open file on common storage (sdcard or internal storage)
 * Updated Lithuanian translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)